product_research/openrouter_cuj/requirements.txt
python-dotenv
openai
django